OKAY SO
PLAYER CLASS

The idea is that resource cards and development cards will not have to be interacted with in the main file, only the player class. So whenever a player decides to buy a development card, one player.command() is issued and the resource cards, development cards, and player are all effected. THIS IS A WORK IN PROGRESS.