+divided all of the files from their .h's to .cpps.
removed the sdlManager. We're going to be using seperate draw functions for different parts of the game. As for the board, think of the map class as having its own sdlmanager built into it. this means that you have to declare all of the images that you want to use in the map in that class, and then call its own internal function.

better words-> for drawing tiles and maps, there are map::apply_surface, map::draw, map::loadImages, map::shutdownImages. this is all part of the map class. if we do a menu screen, i'd imagine it would be its own class, and then require all of these on its own as well (IE menu::draw, menu::apply_surface...)

it builds and is fullscreen, which is goods.